+++ ios/batm/Entities/PreSubmitResponse.swift
@@ -1,6 +1,6 @@
 import Foundation
 
 struct PreSubmitResponse: Equatable {
-  var amount: Double
+  var amount: Decimal
   var address: String
 }
