+++ ios/batm/Entities/PriceChartData.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 struct PriceChartData: Equatable {
-  var price: Double
-  var balance: Double
+  var price: Decimal
+  var balance: Decimal
   var periods: PriceChartPeriods
 }
