+++ ios/batm/Entities/StakeDetails.swift
@@ -3,9 +3,9 @@ import Foundation
 struct StakeDetails: Equatable {
   var exist: Bool
   var unstakeAvailable: Bool
-  var stakedAmount: Double?
-  var rewardsAmount: Double?
-  var rewardsPercent: Double?
+  var stakedAmount: Decimal?
+  var rewardsAmount: Decimal?
+  var rewardsPercent: Decimal?
   var stakedDays: Int?
   var stakingMinDays: Int?
 }
